+++ src/classification_PrecisionRecallCurve.cpp
@@ -7,37 +7,16 @@ using namespace Rcpp;
 //' @method prROC factor
 //' @export
 // [[Rcpp::export(prROC.factor)]]
-Rcpp::DataFrame prROC(const Rcpp::IntegerVector &actual,
-                      const Rcpp::NumericVector &response,
-                      Nullable<bool> micro = R_NilValue,
-                      Rcpp::Nullable<Rcpp::NumericVector> thresholds = R_NilValue,
-                      const bool& na_rm = true) {
+Rcpp::DataFrame PrecisionRecallCurve(const Rcpp::IntegerVector& actual, const Rcpp::NumericVector& response, Rcpp::Nullable<Rcpp::NumericVector> thresholds = R_NilValue) {
+    PRCalculator pr_calculator(actual, response, thresholds);
+    return pr_calculator.calculate();
+}
 
-  /*
-   *  Calculate ROC based
-   * on micro values and
-   * thresholds if passed
-   */
-
-  // 1) default return
-  // value
-  if (micro.isNull()) {
-
-    return _metric_(actual, response, thresholds);
-
-  }
-
-  std::vector<double> empty_numeric;
-  std::vector<int> empty_integer;
-  CharacterVector empty_character;
-
-  Rcpp::DataFrame empty_df = Rcpp::DataFrame::create(
-    Named("threshold") = empty_numeric,
-    Named("level") = empty_integer,
-    Named("label") = empty_character,
-    Named("precision") = empty_numeric,
-    Named("recall") = empty_numeric
-  );
-
-  return empty_df;
+//' @rdname prROC
+//' @method weighted.prROC factor
+//' @export
+// [[Rcpp::export(weighted.prROC.factor)]]
+Rcpp::DataFrame weighted_PrecisionRecallCurve(const Rcpp::IntegerVector& actual, const Rcpp::NumericVector& response, const Rcpp::NumericVector& w, Rcpp::Nullable<Rcpp::NumericVector> thresholds = R_NilValue) {
+    PRCalculator pr_calculator(actual, response, w, thresholds);
+    return pr_calculator.calculate();
 }